requirements.txt
APScheduler==3.6.3
astroid==2.11.2
cachetools==4.2.2
certifi==2021.10.8
dill==0.3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mutagen==1.45.1
platformdirs==2.5.1
pydub==0.25.1
pylint==2.13.5
pyTelegram==0.0.1
python-dotenv==0.21.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
wrapt==1.14.0
